% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowaypipeInAddin.R
\name{twowaypipeInAddin}
\alias{twowaypipeInAddin}
\title{Insert %<>%.}
\usage{
twowaypipeInAddin()
}
\description{
Call this function as an addin to insert \code{ %<>% } (2-way magrittr pipe) at
the cursor position.
}
